requirements.txt
beautifulsoup4
PyGithub
requests
python-dotenv
pytz
pandas